Fix expansion of layout variables inside suffix specifications
The `transient-define-groups' macro saves the parsed suffixes in the
`transient--layout' property, but later when `transient--parse-child'
is used by `transient-define-prefix', that expects the unparsed group
or suffix specifications in the symbol's value cell.  The latter is an
older approach, which was supposed to be replaced when the macro was
introduced.  Because that didn't happen, the macro never worked.

Additionally change `transient--parse-child' to not deal with embedded
symbols, except for quoting them.  It is now `transient--init-child's
responsibility to fill in the indirectly specified suffixes.  In other
words it happens when a prefix is invoked, not when it is defined.
That has the advantage that indirectly defined suffixes can be
modified the same way as directly specified.

Modifying indirectly specified suffixes) now affects all prefixes
that use those suffixes, without having to re-evaluate the prefix
definitions.